# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for flutter app",
        "title": "Hack2023",
        "contact": {},
        "version": "1.0"
    },
    "host": "hack.torbeno.ru",
    "basePath": "/api",
    "paths": {
...
        "/v1/consultation": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "список активных и завершенных консультаций",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "consultation"
                ],
                "summary": "список активных и завершенных консультаций",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Consultations"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "записаться на консультацию",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "consultation"
                ],
                "summary": "записаться на консультацию",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "id надзорного органа",
                        "name": "nadzor_organ_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "id типа контроля",
                        "name": "control_type_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "id темы консультации",
                        "name": "consult_topic_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "id слота с временем и датой консультации",
                        "name": "slot_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "время в формате '03:00'",
                        "name": "time",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "дата в формате '2006-02-01'",
                        "name": "date",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "вопрос в свободной форме",
                        "name": "question",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "нужно ли письменное разъяснение",
                        "name": "is_need_letter",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Consultation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Отменить запись на консультацию",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "consultation"
                ],
                "summary": "Отменить запись на консультацию",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "id консультации которую нужно отменить",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Consultation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Добавление письменного ответа со стороны КНО",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "consultation"
                ],
                "summary": "Добавление письменного ответа со стороны КНО",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "id консультации",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Текст письменного ответа на заданный вопрос",
                        "name": "answer",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Consultation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/v1/user": {
            "get": {
                "description": "Получение данных о пользователе",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Получение данных о пользователе",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Account"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "description": "Сохранить токен приложения пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Сохранить токен приложения пользователя",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Token"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Account": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_kno": {
                    "type": "boolean"
                },
                "nadzor_organ_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.ConsultTopic": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.Consultation": {
            "type": "object",
            "required": [
                "consult_topic_id",
                "control_type_id",
                "nadzor_organ_id",
                "slot_id",
                "time"
            ],
            "properties": {
                "answer": {
                    "type": "string"
                },
                "consult_topic": {
                    "$ref": "#/definitions/model.ConsultTopic"
                },
                "consult_topic_id": {
                    "type": "integer"
                },
                "control_type": {
                    "$ref": "#/definitions/model.ControlType"
                },
                "control_type_id": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_confirmed": {
                    "type": "boolean"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "is_need_letter": {
                    "type": "boolean"
                },
                "nadzor_organ": {
                    "$ref": "#/definitions/model.NadzonOrgan"
                },
                "nadzor_organ_id": {
                    "type": "integer"
                },
                "question": {
                    "type": "string"
                },
                "slot_id": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.Account"
                },
                "user_id": {
                    "type": "integer"
                },
                "video_link": {
                    "type": "string"
                },
                "vks_link": {
                    "type": "string"
                }
            }
        },
        "model.Consultations": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Consultation"
                    }
                },
                "finished": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Consultation"
                    }
                }
            }
        },
        "model.ControlType": {
            "type": "object",
            "properties": {
                "consult_topics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ConsultTopic"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.NadzonOrgan": {
            "type": "object",
            "properties": {
                "control_types": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ControlType"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.ResponseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "model.Token": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter the token with the `Bearer ` prefix, e.g. \"Bearer abcde12345\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}